requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
pypdf==5.1.0
google-genai==1.3.0
pinecone==6.0.0
